move Directive validator and begin adding validation for database connection
