+++ directive/connections.go
@@ -1,10 +1,16 @@
 package directive
 
 import (
+	"fmt"
 	"net/url"
 
 	"github.com/pkg/errors"
 	"github.com/suborbital/reactr/rcap"
+)
+
+const (
+	dbTypeMySQL    = "mysql"
+	dbTypePostgres = "postgresql"
 )
 
 // NATSConnection describes a connection to a NATS server
@@ -69,4 +75,16 @@ func (d *DBConnection) ToRCAPConfig(queries []DBQuery) (*rcap.DatabaseConfig, er
 	}
 
 	return config, nil
+}
+
+func (d *DBConnection) validate() error {
+	if d.Type != dbTypeMySQL && d.Type != dbTypePostgres {
+		return fmt.Errorf("database type %s is invalid, must be 'mysql' or 'postgresql'", d.Type)
+	}
+
+	if d.ConnectionString == "" {
+		return errors.New("database connectionString is empty")
+	}
+
+	return nil
 }
